feat(cangjie_runtime): sync from upstream v1.2.0-alpha.22

+++ runtime/src/Inspector/CjHeapData.h
@@ -53,7 +53,7 @@ public:
     using CjHeapDataStackTraceSerialNumber = CjHeapDataID;
     static constexpr CjHeapDataStackTraceSerialNumber kCjHeapDataNullStackTrace = 0;
     const static size_t alignment = 8;
-    static constexpr size_t HEAP_SIZE_THRESHOLD_4GB = 4ULL * 1024 * 1024 * 1024;
+    static constexpr u8 HEAP_SIZE_THRESHOLD_4GB = 4ULL * 1024 * 1024 * 1024;
     static constexpr u8 NULL_OBJECT_ID = 0xFFFFFFFFFFFFFFFFULL;  // Special ID for null object references
 
     // Encapsulates how cjprof dump records serialize object IDs for the current heap layout.
